+++ exporter/kafkaexporter/kafka_exporter_test.go
@@ -33,35 +33,35 @@ import (
 
 func TestNewExporter_err_version(t *testing.T) {
 	c := Config{ProtocolVersion: "0.0.0", Encoding: defaultEncoding}
-	texp, err := newTracesExporter(c, component.ExporterCreateParams{}, tracesMarshallers())
+	texp, err := newTracesExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, tracesMarshallers())
 	assert.Error(t, err)
 	assert.Nil(t, texp)
 }
 
 func TestNewExporter_err_encoding(t *testing.T) {
 	c := Config{Encoding: "foo"}
-	texp, err := newTracesExporter(c, component.ExporterCreateParams{}, tracesMarshallers())
+	texp, err := newTracesExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, tracesMarshallers())
 	assert.EqualError(t, err, errUnrecognizedEncoding.Error())
 	assert.Nil(t, texp)
 }
 
 func TestNewMetricsExporter_err_version(t *testing.T) {
 	c := Config{ProtocolVersion: "0.0.0", Encoding: defaultEncoding}
-	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{}, metricsMarshallers())
+	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, metricsMarshallers())
 	assert.Error(t, err)
 	assert.Nil(t, mexp)
 }
 
 func TestNewMetricsExporter_err_encoding(t *testing.T) {
 	c := Config{Encoding: "bar"}
-	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{}, metricsMarshallers())
+	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, metricsMarshallers())
 	assert.EqualError(t, err, errUnrecognizedEncoding.Error())
 	assert.Nil(t, mexp)
 }
 
 func TestNewMetricsExporter_err_traces_encoding(t *testing.T) {
 	c := Config{Encoding: "jaeger_proto"}
-	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{}, metricsMarshallers())
+	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, metricsMarshallers())
 	assert.EqualError(t, err, errUnrecognizedEncoding.Error())
 	assert.Nil(t, mexp)
 }
@@ -81,11 +81,11 @@ func TestNewExporter_err_auth_type(t *testing.T) {
 			Full: false,
 		},
 	}
-	texp, err := newTracesExporter(c, component.ExporterCreateParams{}, tracesMarshallers())
+	texp, err := newTracesExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, tracesMarshallers())
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "failed to load TLS config")
 	assert.Nil(t, texp)
-	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{}, metricsMarshallers())
+	mexp, err := newMetricsExporter(c, component.ExporterCreateParams{Logger: zap.NewNop()}, metricsMarshallers())
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "failed to load TLS config")
 	assert.Nil(t, mexp)
@@ -199,7 +199,7 @@ type metricsErrorMarshaller struct {
 	err error
 }
 
-func (e metricsErrorMarshaller) Marshal(traces pdata.Metrics) ([]Message, error) {
+func (e metricsErrorMarshaller) Marshal(_ pdata.Metrics) ([]Message, error) {
 	return nil, e.err
 }
 
@@ -209,7 +209,7 @@ func (e metricsErrorMarshaller) Encoding() string {
 
 var _ TracesMarshaller = (*tracesErrorMarshaller)(nil)
 
-func (e tracesErrorMarshaller) Marshal(traces pdata.Traces) ([]Message, error) {
+func (e tracesErrorMarshaller) Marshal(_ pdata.Traces) ([]Message, error) {
 	return nil, e.err
 }